% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTitle.R
\name{addTitle}
\alias{addTitle}
\title{AddTitle}
\usage{
addTitle(
  object,
  text,
  color = "black",
  fontSize = "20px",
  fontFamily = "Sans",
  leftPosition = 50,
  topPosition = 2
)
}
\description{
https://github.com/Lchiffon/leafletCN/blob/master/R/addTitle.R
}
